Revert "Upgrade to Rails 5.1.4"

+++ db/migrate/20170908160100_add_ip_address_and_user_agent_id_to_event_logs.rb
@@ -5,7 +5,7 @@ class AddIpAddressAndUserAgentIdToEventLogs < ActiveRecord::Migration
     Lhm.cleanup(:run)
     Lhm.change_table :event_logs do |m|
       m.add_column :ip_address, "BIGINT"
-      m.add_column :user_agent_id, "BIGINT"
+      m.add_column :user_agent_id, "INT"
       m.ddl("ALTER TABLE %s ADD CONSTRAINT event_logs_user_agent_id_fk FOREIGN KEY (user_agent_id) REFERENCES user_agents(id)" % m.name)
     end
   end
